// src/app/auth/interfaces/IsgdVwPersonaDTO.ts
export interface IsgdVwPersonaDTO {
    cedula: string;
    nombre: string;
    codPersona: number;
    numRegistroAccion: string;
    nomCategoria: string;
    tipoExperiencia: string;
    nomRegLab: string;
    denoPuesto: string;
    tipoAccion: string;
    fechInicioAccion: Date
    fechFinAccion: Date;
    estado: string;
    explicacionAccion: string;
    nomInstitucion: string;
    amie: string;
    nomCircuito: string;
    nomDistrito: string;
    nomCanton: string;
    nomParroquia: string;
    estadoIE: string;
    estadoSta: string;
    codZona: number;
    celular: string;
    correo: string;
    direccion: string;
}